bsps/shared/grlib: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This is frequently an int or ssize_t variable compared to a uint32_t or
size_t. Sometimes the size_t is from a sizeof() use.

+++ bsps/shared/grlib/can/canbtrs.c
@@ -64,7 +64,7 @@ int grlib_canbtrs_calc_timing(
 		tmp = ((br->divfactor + tseg) * baud);
 		/* Core frequency is always divided by 2 before scaler */
 		sc = core_hz / (2 * tmp);
-		if (sc <= 0 || sc > br->max_scaler)
+		if (sc <= 0 || (unsigned int)sc > br->max_scaler)
 			continue;
 		if (br->has_bpr &&
 		    (((sc > 256 * 1) && (sc <= 256 * 2) && (sc & 0x1)) ||
